# tremor-guard-backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TremorGuard Backend",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/reports": {
      "get": {
        "tags": [
          "reports"
        ],
        "summary": "List Reports",
        "operationId": "list_reports_v1_reports_get",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "reports"
        ],
        "summary": "Create Report",
        "operationId": "create_report_v1_reports_post",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateReportRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateReportResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateReportRequest": {
        "properties": {
          "report_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Report Date"
          },
          "report_type": {
            "type": "string",
            "title": "Report Type",
            "default": "周度病情评估摘要"
          }
        },
        "type": "object",
        "title": "CreateReportRequest"
      },
      "CreateReportResponse": {
        "properties": {
          "report": {
            "$ref": "#/components/schemas/ReportSummaryDTO"
          }
        },
        "type": "object",
        "required": [
          "report"
        ],
        "title": "CreateReportResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ReportListResponse": {
        "properties": {
          "report_summaries": {
            "items": {
              "$ref": "#/components/schemas/ReportSummaryDTO"
            },
            "type": "array",
            "title": "Report Summaries"
          }
        },
        "type": "object",
        "required": [
          "report_summaries"
        ],
        "title": "ReportListResponse"
      },
      "ReportSummaryDTO": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "date": {
            "type": "string",
            "format": "date",
            "title": "Date"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "size": {
            "type": "string",
            "title": "Size"
          },
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "id",
          "date",
          "type",
          "size",
          "status"
        ],
        "title": "ReportSummaryDTO"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}